CS50 Sentiments
My solution to Problem 6 from CS50 Harvard course. Web application made
in Flask that you give a tweeter account and the program, with help
from Natural Language Toolkit library, analyzes tweets and compare
words to a dictionary of positive and negative words. Later draws a pie
chart with the statistics of how “positive and negative” the user is
based on his tweets.